Get SD Card Working
The virtual SD card was not setup, because the "on nonencrypted" trigger
that launches the "late_start" action class was not being called. The
sdcard service is part of the late_start action class.

A mount_all command is necessary to trigger it which requires a fstab.mako
file. Create fstab.mako with one entry for /system in it, so that
nonencrypted trigger is fired. This converted entry is probably missing
some options in the conversion like e2fsck. The rest of the mount commands
from init.mako.rc need to be converted as well.

Change-Id: Ifdf489a4edb323ce05ca8b2df18892cfcf72043c