mako: Update wpa_supplicant.conf to support concurrency
In JellyBean the wpa_supplicant.conf file is used as a template for
creating both the wlan0 and p2p0 configuration files.  In order to
support the simplistic way the template is modified by wifi.c we
remove all instances of the string "ctrl_interface" other than the one
that should be modified, and we explicitly modify the assigned value
to denote that it will be modified by wifi.c.

Change-Id: I484a5beeb711b35668693f7cbaf6f91860a3a83c