mako: bump up lowest brightness setting
b/7128723

This increases the lowest brightness value, so that the screen is more
legible in a dim lit room. It is probably too bright in pitch dark if
one had to choose, but it seems to be a bit better in any scenario where
there is a little ambient light.

Change-Id: I3a413ef5c26aa9f49cdfe376df022a1e0c957027
Signed-off-by: Iliyan Malchev <malchev@google.com>