mako: relocate ramdisk to avoid overlap with kernel
An upcoming change will statically link the Prima WLAN Driver with the
kernel.  The current observation is that when this occurs, the kernel
will increase in size to the point that it will overlap with the
ramdisk, preventing the ramdisk from being mounted.  In anticipation
of this, increase the ramdisk_offset so that the kernel with the
statically-linked Prima WLAN Driver will not overlap the ramdisk.

Change-Id: If5a0ffe67f6ba044690d602d5f04879a983a52b7