mako: increase first brightness range bound
Adjust the first range's upper bound from 1 to 3. Sometimes in the
morning when there is some light in the room it is too drastic to bump
up to the next stage of brightness. In the mornings, the power manager
and twilight service seem to use the configured values, but at night,
it tends to reduce the configured values by quite a bit.

Change-Id: Ia744444e60f10f044c70f1918616effce218ea8d