Vibration tuning adapted to mako from gnexusparts
This package adds a vibration tuning slider in advanced settings to adjust the power of the vibration on the device.
Write permission was set in order to change the amp value. This is the gnexusparts source modified to work for mako.

The max setting in the kernel is 88, this slider will go to 100. I debated maxing the slider out at 88 instead of allowing it to go
to an untested level of 100 but other examples of tuning allow user control to go past recommend settings so I submitted it this way.
This is an easy thing to change if its deemed unsafe to release as is.

Change-Id: I93c03954fe275ed9aa7dcd32c5a4226b383504b6