Supply the full path to sysfs_rmnet.
On my device, /sys/class/android_usb/f_rmnet_smd_sdio is a
symlink to
../../devices/virtual/android_usb/android0/f_rmnet_smd_sdio .
When the recursive restorecon is performed on /sys, it ignores
symlinks and only processes actual files.

This causes problems for the
/sys/class/android_usb/f_rmnet_smd_sdio/transport entry in
file_contexts. After you resolve symlinks, this path
expands out to:
/sys/devices/virtual/android_usb/android0/f_rmnet_smd_sdio/transport
which doesn't match the regular expression in file_contexts.

To fix this problem, we add an entry for the fully resolved
path to file_contexts.

This addresses the following denial:

<5>[   17.380405] type=1400 audit(1383244856.165:5): avc:  denied  { write } for  pid=182 comm="bridgemgrd" name="transport" dev="sysfs" ino=13405 scontext=u:r:bridge:s0 tcontext=u:object_r:sysfs:s0 tclass=file

Change-Id: If3a0d90b1f281f2154668a2eaa7d4a67b8e39e75