Move audio_firmware_file and /data/misc/audio entry to core sepolicy.
file_contexts uses regexes, not a globs, so use (/.*)? rather than /*
to match the directory and anything beneath it.

Since /data/misc/audio is not device-specific, move it to core sepolicy.

Consider renaming this type in the future to audio_data_file, but that
is left to a separate change as it will require a restorecon_recursive
on mako.

Change-Id: Ib8c96ab9e19d34e8e34a4c859528345763be4906
Signed-off-by: Stephen Smalley <sds@tycho.nsa.gov>