Change camera_calibration_file to camera_data_file in policy.
The camera_calibration_file label is used for /data/fdAlbum
but that should be changed to camera_data_file. There is
an individual restorecon in init.mako.rc to relabel this
on boot. A change to the file_contexts has also been
made which will cause the hash of the file to change and
trigger the restorecon during an OTA.

Change-Id: I7391a137f0e050b3f27c806e93993118860641b5
Signed-off-by: rpcraig <rpcraig@tycho.ncsc.mil>