Add LTE to "config_radio_access_family" string
The Nexus 4 is capable of LTE data using the correct modem image.  The ability to switch to LTE was supported in CM12, CM11, and CM10 (at a minimum).  Simply adding this feature back.

Change was tested on an unofficial build of the 030816 mako nightly.  Switching radio to "LTE Only", and then to "LTE/GSM (auto)" results in LTE data becoming enabled.